orchestrator/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pyjwt==2.8.0
pydantic==1.10.15
